chore: refactor UserSignup controller
