+++ pkg/controller/usersignup/usersignup_controller_test.go
@@ -6,17 +6,13 @@ import (
 	"encoding/hex"
 	"errors"
 	"fmt"
-	"os"
 	"strconv"
 	"testing"
 
 	"github.com/codeready-toolchain/host-operator/pkg/counter"
 	"github.com/codeready-toolchain/host-operator/pkg/templates/nstemplatetiers"
-	. "github.com/codeready-toolchain/host-operator/test"
-	"github.com/operator-framework/operator-sdk/pkg/k8sutil"
-	"sigs.k8s.io/controller-runtime/pkg/event"
-	"sigs.k8s.io/controller-runtime/pkg/handler"
 
+	. "github.com/codeready-toolchain/host-operator/test"
 	murtest "github.com/codeready-toolchain/toolchain-common/pkg/test/masteruserrecord"
 
 	"github.com/codeready-toolchain/api/pkg/apis/toolchain/v1alpha1"
@@ -242,11 +238,10 @@ func TestUserSignupWithMissingEmailLabelFails(t *testing.T) {
 	defer clearMemberClusters(r.client)
 
 	// when
-	res, err := r.Reconcile(req)
+	_, err := r.Reconcile(req)
 
 	// then
-	require.NoError(t, err)
-	require.Equal(t, reconcile.Result{}, res)
+	require.Error(t, err)
 
 	// Lookup the user signup again
 	err = r.client.Get(context.TODO(), types.NamespacedName{Name: userSignup.Name, Namespace: req.Namespace}, userSignup)
@@ -254,9 +249,10 @@ func TestUserSignupWithMissingEmailLabelFails(t *testing.T) {
 
 	test.AssertConditionsMatch(t, userSignup.Status.Conditions,
 		v1alpha1.Condition{
-			Type:   v1alpha1.UserSignupComplete,
-			Status: v1.ConditionFalse,
-			Reason: "MissingUserEmailAnnotation",
+			Type:    v1alpha1.UserSignupComplete,
+			Status:  v1.ConditionFalse,
+			Reason:  "MissingUserEmailAnnotation",
+			Message: "missing annotation at usersignup",
 		})
 	AssertThatCounterHas(t, 1)
 }
@@ -289,11 +285,10 @@ func TestUserSignupWithInvalidEmailHashLabelFails(t *testing.T) {
 	defer clearMemberClusters(r.client)
 
 	// when
-	res, err := r.Reconcile(req)
+	_, err := r.Reconcile(req)
 
 	// then
-	require.NoError(t, err)
-	require.Equal(t, reconcile.Result{}, res)
+	require.Error(t, err)
 
 	// Lookup the user signup again
 	err = r.client.Get(context.TODO(), types.NamespacedName{Name: userSignup.Name, Namespace: req.Namespace}, userSignup)
@@ -301,9 +296,10 @@ func TestUserSignupWithInvalidEmailHashLabelFails(t *testing.T) {
 
 	test.AssertConditionsMatch(t, userSignup.Status.Conditions,
 		v1alpha1.Condition{
-			Type:   v1alpha1.UserSignupComplete,
-			Status: v1.ConditionFalse,
-			Reason: "InvalidEmailHashLabel",
+			Type:    v1alpha1.UserSignupComplete,
+			Status:  v1.ConditionFalse,
+			Reason:  "InvalidEmailHashLabel",
+			Message: "hash is invalid",
 		})
 	AssertThatCounterHas(t, 1)
 }
@@ -333,11 +329,10 @@ func TestUserSignupWithMissingEmailHashLabelFails(t *testing.T) {
 	defer clearMemberClusters(r.client)
 
 	// when
-	res, err := r.Reconcile(req)
+	_, err := r.Reconcile(req)
 
 	// then
-	require.NoError(t, err)
-	require.Equal(t, reconcile.Result{}, res)
+	require.Error(t, err)
 
 	// Lookup the user signup again
 	err = r.client.Get(context.TODO(), types.NamespacedName{Name: userSignup.Name, Namespace: req.Namespace}, userSignup)
@@ -345,9 +340,10 @@ func TestUserSignupWithMissingEmailHashLabelFails(t *testing.T) {
 
 	test.AssertConditionsMatch(t, userSignup.Status.Conditions,
 		v1alpha1.Condition{
-			Type:   v1alpha1.UserSignupComplete,
-			Status: v1.ConditionFalse,
-			Reason: "MissingEmailHashLabel",
+			Type:    v1alpha1.UserSignupComplete,
+			Status:  v1.ConditionFalse,
+			Reason:  "MissingEmailHashLabel",
+			Message: "missing label at usersignup",
 		})
 	AssertThatCounterHas(t, 1)
 }
@@ -377,6 +373,11 @@ func TestUserSignupFailedMissingNSTemplateTier(t *testing.T) {
 	require.NoError(t, err)
 	t.Logf("usersignup status: %+v", userSignup.Status)
 	test.AssertConditionsMatch(t, userSignup.Status.Conditions,
+		v1alpha1.Condition{
+			Type:   v1alpha1.UserSignupApproved,
+			Status: v1.ConditionTrue,
+			Reason: "ApprovedAutomatically",
+		},
 		v1alpha1.Condition{
 			Type:    v1alpha1.UserSignupComplete,
 			Status:  v1.ConditionFalse,
@@ -1718,197 +1719,6 @@ func TestUserSignupNoMembersAvailableFails(t *testing.T) {
 	AssertThatCounterHas(t, 1)
 }
 
-func TestBannedUserToUserSignupMapper(t *testing.T) {
-	// given
-	bannedUser := &toolchainv1alpha1.BannedUser{
-		ObjectMeta: metav1.ObjectMeta{
-			Labels: map[string]string{
-				toolchainv1alpha1.BannedUserEmailHashLabelKey: "fd2addbd8d82f0d2dc088fa122377eaa",
-			},
-		},
-		Spec: toolchainv1alpha1.BannedUserSpec{
-			Email: "foo@redhat.com",
-		},
-	}
-
-	t.Run("test BannedUserToUserSignupMapper maps correctly", func(t *testing.T) {
-		userSignup := &v1alpha1.UserSignup{
-			ObjectMeta: newObjectMeta("", "foo@redhat.com"),
-			Spec: v1alpha1.UserSignupSpec{
-				Username: "foo@redhat.com",
-			},
-		}
-
-		userSignup2 := &v1alpha1.UserSignup{
-			ObjectMeta: newObjectMeta("", "alice.mayweather.doe@redhat.com"),
-			Spec: v1alpha1.UserSignupSpec{
-				Username: "alice.mayweather.doe@redhat.com",
-			},
-		}
-
-		c := test.NewFakeClient(t, userSignup, userSignup2)
-
-		mapper := &BannedUserToUserSignupMapper{
-			client: c,
-		}
-
-		// This is required for the mapper to function
-		os.Setenv(k8sutil.WatchNamespaceEnvVar, operatorNamespace)
-		defer os.Unsetenv(k8sutil.WatchNamespaceEnvVar)
-
-		req := mapper.Map(handler.MapObject{
-			Object: bannedUser,
-		})
-
-		require.Len(t, req, 1)
-		require.Equal(t, types.NamespacedName{
-			Namespace: userSignup.Namespace,
-			Name:      userSignup.Name,
-		}, req[0].NamespacedName)
-	})
-
-	t.Run("test BannedUserToUserSignupMapper returns nil when client list fails", func(t *testing.T) {
-		c := test.NewFakeClient(t)
-		c.MockList = func(ctx context.Context, list runtime.Object, opts ...client.ListOption) error {
-			return errors.New("err happened")
-		}
-
-		mapper := &BannedUserToUserSignupMapper{
-			client: c,
-		}
-		req := mapper.Map(handler.MapObject{
-			Object: bannedUser,
-		})
-
-		require.Nil(t, req)
-	})
-
-	t.Run("test BannedUserToUserSignupMapper returns nil when watch namespace not set ", func(t *testing.T) {
-		c := test.NewFakeClient(t)
-
-		mapper := &BannedUserToUserSignupMapper{
-			client: c,
-		}
-		req := mapper.Map(handler.MapObject{
-			Object: bannedUser,
-		})
-
-		require.Nil(t, req)
-	})
-}
-
-func TestUserSignupChangedPredicate(t *testing.T) {
-	// when
-	pred := &UserSignupChangedPredicate{}
-	userSignupName := uuid.NewV4().String()
-
-	userSignupOld := &v1alpha1.UserSignup{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      userSignupName,
-			Namespace: operatorNamespace,
-			Annotations: map[string]string{
-				toolchainv1alpha1.UserSignupUserEmailAnnotationKey: "foo@redhat.com",
-			},
-			Labels: map[string]string{
-				toolchainv1alpha1.UserSignupUserEmailHashLabelKey: "fd2addbd8d82f0d2dc088fa122377eaa",
-			},
-			Generation: 1,
-		},
-		Spec: v1alpha1.UserSignupSpec{
-			Username: "foo@redhat.com",
-		},
-	}
-
-	userSignupNewNotChanged := &v1alpha1.UserSignup{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      userSignupName,
-			Namespace: operatorNamespace,
-			Annotations: map[string]string{
-				toolchainv1alpha1.UserSignupUserEmailAnnotationKey: "foo@redhat.com",
-			},
-			Labels: map[string]string{
-				toolchainv1alpha1.UserSignupUserEmailHashLabelKey: "fd2addbd8d82f0d2dc088fa122377eaa",
-			},
-			Generation: 1,
-		},
-		Spec: v1alpha1.UserSignupSpec{
-			Username: "alice.mayweather.doe@redhat.com",
-		},
-	}
-
-	userSignupNewChanged := &v1alpha1.UserSignup{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      userSignupName,
-			Namespace: operatorNamespace,
-			Annotations: map[string]string{
-				toolchainv1alpha1.UserSignupUserEmailAnnotationKey: "alice.mayweather.doe@redhat.com",
-			},
-			Labels: map[string]string{
-				toolchainv1alpha1.UserSignupUserEmailHashLabelKey: "747a250430df0c7976bf2363ebb4014a",
-			},
-			Generation: 2,
-		},
-		Spec: v1alpha1.UserSignupSpec{
-			Username: "alice.mayweather.doe@redhat.com",
-		},
-	}
-
-	t.Run("test UserSignupChangedPredicate returns false when MetaOld not set", func(t *testing.T) {
-		e := event.UpdateEvent{
-			MetaOld:   nil,
-			ObjectOld: userSignupOld,
-			MetaNew:   userSignupNewNotChanged.ObjectMeta.GetObjectMeta(),
-			ObjectNew: userSignupNewNotChanged,
-		}
-		require.False(t, pred.Update(e))
-	})
-	t.Run("test UserSignupChangedPredicate returns false when ObjectOld not set", func(t *testing.T) {
-		e := event.UpdateEvent{
-			MetaOld:   userSignupOld.ObjectMeta.GetObjectMeta(),
-			ObjectOld: nil,
-			MetaNew:   userSignupNewNotChanged.ObjectMeta.GetObjectMeta(),
-			ObjectNew: userSignupNewNotChanged,
-		}
-		require.False(t, pred.Update(e))
-	})
-	t.Run("test UserSignupChangedPredicate returns false when ObjectNew not set", func(t *testing.T) {
-		e := event.UpdateEvent{
-			MetaOld:   userSignupOld.ObjectMeta.GetObjectMeta(),
-			ObjectOld: userSignupOld,
-			MetaNew:   userSignupNewNotChanged.ObjectMeta.GetObjectMeta(),
-			ObjectNew: nil,
-		}
-		require.False(t, pred.Update(e))
-	})
-	t.Run("test UserSignupChangedPredicate returns false when MetaNew not set", func(t *testing.T) {
-		e := event.UpdateEvent{
-			MetaOld:   userSignupOld.ObjectMeta.GetObjectMeta(),
-			ObjectOld: userSignupOld,
-			MetaNew:   nil,
-			ObjectNew: userSignupNewNotChanged,
-		}
-		require.False(t, pred.Update(e))
-	})
-	t.Run("test UserSignupChangedPredicate returns false when generation unchanged and annoations unchanged", func(t *testing.T) {
-		e := event.UpdateEvent{
-			MetaOld:   userSignupOld.ObjectMeta.GetObjectMeta(),
-			ObjectOld: userSignupOld,
-			MetaNew:   userSignupNewNotChanged.ObjectMeta.GetObjectMeta(),
-			ObjectNew: userSignupNewNotChanged,
-		}
-		require.False(t, pred.Update(e))
-	})
-	t.Run("test UserSignupChangedPredicate returns true when generation changed", func(t *testing.T) {
-		e := event.UpdateEvent{
-			MetaOld:   userSignupOld.ObjectMeta.GetObjectMeta(),
-			ObjectOld: userSignupOld,
-			MetaNew:   userSignupNewChanged.ObjectMeta.GetObjectMeta(),
-			ObjectNew: userSignupNewChanged,
-		}
-		require.True(t, pred.Update(e))
-	})
-}
-
 func prepareReconcile(t *testing.T, name string, initObjs ...runtime.Object) (*ReconcileUserSignup, reconcile.Request, *test.FakeClient) {
 	s := scheme.Scheme
 	err := apis.AddToScheme(s)
@@ -1930,7 +1740,9 @@ func prepareReconcile(t *testing.T, name string, initObjs ...runtime.Object) (*R
 	client := test.NewFakeClient(t, initObjs...)
 
 	r := &ReconcileUserSignup{
-		client: client,
+		statusUpdater: &statusUpdater{
+			client: client,
+		},
 		scheme: s,
 	}
 	return r, newReconcileRequest(name), client
